+++ mne/io/base.py
@@ -1438,13 +1438,16 @@ class _BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
 
         remove_dc : bool
             If True remove DC component when plotting data.
-        order : 'type' | 'original' | 'selection' | 'position' | array of int
+        order : str | array of int
             Order in which to plot data. 'type' groups by channel type,
             'original' plots in the order of ch_names, 'selection' uses
-            Elekta's channel groupings (only works for Neuromag data) and
-            'positions' groups the channels by the positions of the sensors.
-            If array, the order is determined by the indices in the array.
-            Defaults to 'type'.
+            Elekta's channel groupings (only works for Neuromag data),
+            'position' groups the channels by the positions of the sensors.
+            'selection' and 'position' modes allow custom selections by using
+            lasso selector on the topomap. Pressing ``ctrl`` key while
+            selecting allows appending to the current selection. If array, only
+            the channels in the array are plotted in the given order. Defaults
+            to 'type'.
         show_options : bool
             If True, a dialog for options related to projection is shown.
         title : str | None