+++ mne/viz/utils.py
@@ -17,6 +17,7 @@ import webbrowser
 import tempfile
 import numpy as np
 from copy import deepcopy
+from distutils.version import LooseVersion
 
 from ..channels.layout import _auto_topomap_coords
 from ..channels.channels import _contains_ch_type
@@ -584,10 +585,11 @@ def _radio_clicked(label, params):
     params['fig_selection'].radio._active_idx = idx
     channels = params['selections'][label]
     ax_topo = params['fig_selection'].get_axes()[1]
-    for type in ('mag', 'grad', 'eeg', 'seeg'):
-        if type in params['types']:
-            types = np.where(np.array(params['types']) == type)[0]
-            break
+    types = np.array([], dtype=int)
+    for this_type in ('mag', 'grad', 'eeg', 'seeg', 'ecog'):
+        if this_type in params['types']:
+            types = np.concatenate(
+                [types, np.where(np.array(params['types']) == this_type)[0]])
     colors = np.zeros((len(types), 4))  # alpha = 0 by default
     locs3d = np.array([ch['loc'][:3] for ch in params['info']['chs']])
     x, y, z = locs3d.T
@@ -606,7 +608,8 @@ def _radio_clicked(label, params):
     for line in params['lines'][n_channels:]:  # To remove lines from view.
         line.set_xdata([])
         line.set_ydata([])
-    _setup_browser_offsets(params, n_channels)
+    if n_channels > 0:  # Can be 0 with lasso selector.
+        _setup_browser_offsets(params, n_channels)
     params['plot_fun']()
 
 
@@ -772,6 +775,8 @@ def _find_channel_idx(ch_name, params):
     offset = 0
     labels = [l._text for l in params['fig_selection'].radio.labels]
     for label in labels:
+        if label == 'Custom':
+            continue  # Custom selection not included as it shifts the indices.
         selection = params['selections'][label]
         hits = np.where(np.array(params['raw'].ch_names)[selection] == ch_name)
         for idx in hits[0]:
@@ -1072,7 +1077,8 @@ def _process_times(inst, times, n_peaks=None, few=False):
 
 
 def plot_sensors(info, kind='topomap', ch_type=None, title=None,
-                 show_names=False, ch_groups=None, axes=None, show=True):
+                 show_names=False, ch_groups=None, axes=None, block=False,
+                 show=True):
     """Plot sensors positions.
 
     Parameters
@@ -1080,11 +1086,17 @@ def plot_sensors(info, kind='topomap', ch_type=None, title=None,
     info : Instance of Info
         Info structure containing the channel locations.
     kind : str
-        Whether to plot the sensors as 3d or as topomap. Available options
-        'topomap', '3d'. Defaults to 'topomap'.
-    ch_type : 'mag' | 'grad' | 'eeg' | 'seeg' | None
-        The channel type to plot. If None, then channels are chosen in the
-        order given above.
+        Whether to plot the sensors as 3d, topomap or as an interactive
+        sensor selection dialog. Available options 'topomap', '3d', 'select'.
+        If 'select', a set of channels can be selected interactively by using
+        lasso selector or clicking while holding control key. The selected
+        channels are returned along with the figure instance. Defaults to
+        'topomap'.
+    ch_type : None | str
+        The channel type to plot. Available options 'mag', 'grad', 'eeg',
+        'seeg', 'ecog', 'all'. If ``'all'``, all the available mag, grad, eeg,
+        seeg and ecog channels are plotted. If None (default), then channels
+        are chosen in the order given above.
     title : str | None
         Title for the figure. If None (default), equals to
         ``'Sensor positions (%s)' % ch_type``.
@@ -1104,6 +1116,12 @@ def plot_sensors(info, kind='topomap', ch_type=None, title=None,
 
         .. versionadded:: 0.13.0
 
+    block : bool
+        Whether to halt program execution until the figure is closed. Defaults
+        to False.
+
+        .. versionadded:: 0.13.0
+
     show : bool
         Show figure if True. Defaults to True.
 
@@ -1111,6 +1129,8 @@ def plot_sensors(info, kind='topomap', ch_type=None, title=None,
     -------
     fig : instance of matplotlib figure
         Figure containing the sensor topography.
+    selection : list
+        A list of selected channels. Only returned if ``kind=='select'``.
 
     See Also
     --------
@@ -1126,21 +1146,29 @@ def plot_sensors(info, kind='topomap', ch_type=None, title=None,
 
     """
     from .evoked import _rgb
-    if kind not in ['topomap', '3d']:
-        raise ValueError("Kind must be 'topomap' or '3d'.")
+    if kind not in ['topomap', '3d', 'select']:
+        raise ValueError("Kind must be 'topomap', '3d' or 'select'. Got %s." %
+                         kind)
     if not isinstance(info, Info):
         raise TypeError('info must be an instance of Info not %s' % type(info))
     ch_indices = channel_indices_by_type(info)
-    allowed_types = ['mag', 'grad', 'eeg', 'seeg']
+    allowed_types = ['mag', 'grad', 'eeg', 'seeg', 'ecog']
     if ch_type is None:
         for this_type in allowed_types:
             if _contains_ch_type(info, this_type):
                 ch_type = this_type
                 break
-    elif ch_type not in allowed_types:
+        picks = ch_indices[ch_type]
+    elif ch_type == 'all':
+        picks = list()
+        for this_type in allowed_types:
+            picks += ch_indices[this_type]
+    elif ch_type in allowed_types:
+        picks = ch_indices[ch_type]
+    else:
         raise ValueError("ch_type must be one of %s not %s!" % (allowed_types,
                                                                 ch_type))
-    picks = ch_indices[ch_type]
+
     if len(picks) == 0:
         raise ValueError('Could not find any channels of type %s.' % ch_type)
 
@@ -1184,19 +1212,29 @@ def plot_sensors(info, kind='topomap', ch_type=None, title=None,
                 if pick in value:
                     colors[pick_idx] = color_vals[ind]
                     break
-    if kind == 'topomap':
+    if kind in ('topomap', 'select'):
         pos = _auto_topomap_coords(info, picks, True)
 
     title = 'Sensor positions (%s)' % ch_type if title is None else title
     fig = _plot_sensors(pos, colors, bads, ch_names, title, show_names, axes,
-                        show)
+                        show, kind == 'select', block=block)
+    if kind == 'select':
+        return fig, fig.lasso.selection
     return fig
 
 
-def _onpick_sensor(event, fig, ax, pos, ch_names):
+def _onpick_sensor(event, fig, ax, pos, ch_names, show_names):
     """Callback for picked channel in plot_sensors."""
+    if event.mouseevent.key == 'control' and fig.lasso is not None:
+        for ind in event.ind:
+            fig.lasso.select_one(ind)
+
+        return
+    if show_names:
+        return  # channel names already visible
     ind = event.ind[0]  # Just take the first sensor.
     ch_name = ch_names[ind]
+
     this_pos = pos[ind]
 
     # XXX: Bug in matplotlib won't allow setting the position of existing
@@ -1209,7 +1247,12 @@ def _onpick_sensor(event, fig, ax, pos, ch_names):
     fig.canvas.draw()
 
 
-def _plot_sensors(pos, colors, bads, ch_names, title, show_names, ax, show):
+def _close_event(event, fig):
+    fig.lasso.disconnect()
+
+
+def _plot_sensors(pos, colors, bads, ch_names, title, show_names, ax, show,
+                  select, block):
     """Helper function for plotting sensors."""
     import matplotlib.pyplot as plt
     from mpl_toolkits.mplot3d import Axes3D
@@ -1230,6 +1273,7 @@ def _plot_sensors(pos, colors, bads, ch_names, title, show_names, ax, show):
         ax.text(0, 0, 0, '', zorder=1)
         ax.scatter(pos[:, 0], pos[:, 1], pos[:, 2], picker=True, c=colors,
                    s=75, edgecolor=edgecolors, linewidth=2)
+
         ax.azim = 90
         ax.elev = 0
     else:
@@ -1240,9 +1284,13 @@ def _plot_sensors(pos, colors, bads, ch_names, title, show_names, ax, show):
                             hspace=None)
         pos, outlines = _check_outlines(pos, 'head')
         _draw_outlines(ax, outlines)
-        ax.scatter(pos[:, 0], pos[:, 1], picker=True, c=colors, s=75,
-                   edgecolor=edgecolors, linewidth=2)
 
+        pts = ax.scatter(pos[:, 0], pos[:, 1], picker=True, c=colors, s=75,
+                         edgecolor=edgecolors, linewidth=2)
+        if select:
+            fig.lasso = SelectFromCollection(ax, pts, ch_names)
+
+    connect_picker = True
     if show_names:
         for idx in range(len(pos)):
             this_pos = pos[idx]
@@ -1250,13 +1298,16 @@ def _plot_sensors(pos, colors, bads, ch_names, title, show_names, ax, show):
                 ax.text(this_pos[0], this_pos[1], this_pos[2], ch_names[idx])
             else:
                 ax.text(this_pos[0], this_pos[1], ch_names[idx])
-    else:
+        connect_picker = select
+    if connect_picker:
         picker = partial(_onpick_sensor, fig=fig, ax=ax, pos=pos,
-                         ch_names=ch_names)
+                         ch_names=ch_names, show_names=show_names)
         fig.canvas.mpl_connect('pick_event', picker)
 
     fig.suptitle(title)
-    plt_show(show)
+    closed = partial(_close_event, fig=fig)
+    fig.canvas.mpl_connect('close_event', closed)
+    plt_show(show, block=block)
     return fig
 
 
@@ -1422,3 +1473,100 @@ class DraggableColorbar(object):
         self.cbar.draw_all()
         self.mappable.set_norm(self.cbar.norm)
         self.cbar.patch.figure.canvas.draw()
+
+
+class SelectFromCollection(object):
+    """Select channels from a matplotlib collection using `LassoSelector`.
+
+    Selected channels are saved in the ``selection`` attribute. This tool
+    highlights selected points by fading other points out (i.e., reducing their
+    alpha values).
+
+    Notes:
+    This tool selects collection objects based on their *origins*
+    (i.e., `offsets`). Emits mpl event 'lasso_event' when selection is ready.
+
+    Parameters
+    ----------
+    ax : Instance of Axes
+        Axes to interact with.
+
+    collection : Instance of matplotlib collection
+        Collection you want to select from.
+
+    alpha_other : 0 <= float <= 1
+        To highlight a selection, this tool sets all selected points to an
+        alpha value of 1 and non-selected points to `alpha_other`.
+        Defaults to 0.3.
+    """
+
+    def __init__(self, ax, collection, ch_names, alpha_other=0.3):
+        import matplotlib as mpl
+        if LooseVersion(mpl.__version__) < LooseVersion('1.2.1'):
+            raise ImportError('Interactive selection not possible for '
+                              'matplotlib versions < 1.2.1. Upgrade '
+                              'matplotlib.')
+        from matplotlib.widgets import LassoSelector
+        self.canvas = ax.figure.canvas
+        self.collection = collection
+        self.ch_names = ch_names
+        self.alpha_other = alpha_other
+
+        self.xys = collection.get_offsets()
+        self.Npts = len(self.xys)
+
+        # Ensure that we have separate colors for each object
+        self.fc = collection.get_facecolors()
+        if len(self.fc) == 0:
+            raise ValueError('Collection must have a facecolor')
+        elif len(self.fc) == 1:
+            self.fc = np.tile(self.fc, self.Npts).reshape(self.Npts, -1)
+        self.fc[:, -1] = self.alpha_other  # deselect in the beginning
+
+        self.lasso = LassoSelector(ax, onselect=self.on_select,
+                                   lineprops={'color': 'red', 'linewidth': .5})
+        self.selection = list()
+
+    def on_select(self, verts):
+        """Callback for selecting a subset from the collection."""
+        from matplotlib.path import Path
+        if len(verts) <= 3:  # Seems to be a good way to exclude single clicks.
+            return
+
+        path = Path(verts)
+        inds = np.nonzero([path.contains_point(xy) for xy in self.xys])[0]
+        if self.canvas._key == 'control':  # Appending selection.
+            sels = [np.where(self.ch_names == c)[0][0] for c in self.selection]
+            inters = set(inds) - set(sels)
+            inds = list(inters.union(set(sels) - set(inds)))
+
+        while len(self.selection) > 0:
+            self.selection.pop(0)
+        self.selection.extend(self.ch_names[inds])
+        self.fc[:, -1] = self.alpha_other
+        self.fc[inds, -1] = 1
+        self.collection.set_facecolors(self.fc)
+        self.canvas.draw_idle()
+        self.canvas.callbacks.process('lasso_event')
+
+    def select_one(self, ind):
+        """Helper for selecting/deselecting one sensor."""
+        ch_name = self.ch_names[ind]
+        if ch_name in self.selection:
+            sel_ind = self.selection.index(ch_name)
+            self.selection.pop(sel_ind)
+            this_alpha = self.alpha_other
+        else:
+            self.selection.append(ch_name)
+            this_alpha = 1
+        self.fc[ind, -1] = this_alpha
+        self.collection.set_facecolors(self.fc)
+        self.canvas.draw_idle()
+        self.canvas.callbacks.process('lasso_event')
+
+    def disconnect(self):
+        """Method for disconnecting the lasso selector."""
+        self.lasso.disconnect_events()
+        self.fc[:, -1] = 1
+        self.collection.set_facecolors(self.fc)
+        self.canvas.draw_idle()